added friends leaderboard ahhh

+++ models/scoreboard-model.js
@@ -1,23 +1,27 @@
 const mongoose = require("mongoose");
+const dbcommons = require("./dbcommons");
 
 const scoreboardSchema = new mongoose.Schema({
     playerId: {type: String, required: true, unique: true},
     username: {type: String, required: true},
     highscore: {type: Number, required: true},
-    gamesPlayed: {type: Number, required: true}
+    gamesPlayed: {type: Number, required: true},
+    lastPlayed: {type: Date, default: Date.now}
 })
 
 const Scoreboard = mongoose.model('Scoreboard', scoreboardSchema);
 
 exports.getTopTen = function() {
+    if (!dbcommons.isDBConnected()) throw dbcommons.databaseError;
     return Scoreboard.find().sort({ highscore: -1 }).limit(10);
 };
 
 exports.upsertScore = function(playerId, username, newScore) {
+    if (!dbcommons.isDBConnected()) throw dbcommons.databaseError;
     return Scoreboard.findOneAndUpdate(
         { playerId: playerId },
         {
-            $set: { username: username },
+            $set: { username: username, lastPlayed: new Date() },
             $max: { highscore: newScore },
             $inc: { gamesPlayed: 1 }
         },
@@ -26,5 +30,11 @@ exports.upsertScore = function(playerId, username, newScore) {
 };
 
 exports.deleteScore = function(playerId) {
+    if (!dbcommons.isDBConnected()) throw dbcommons.databaseError;
     return Scoreboard.deleteOne({ playerId: playerId });
+};
+
+exports.getFriendScores = function(playerIds) {
+    if (!dbcommons.isDBConnected()) throw dbcommons.databaseError;
+    return Scoreboard.find({playerId: {$in: playerIds}}).sort({highscore: -1});
 };
